+++ src/Components/IndexScene/IndexScene.tsx
@@ -24,7 +24,6 @@ import {
   sceneGraph,
 } from '@grafana/scenes';
 import { Text, useStyles2 } from '@grafana/ui';
-import { ServiceScene } from 'Components/MainScenes/ServiceScene';
 import {
   VAR_DATASOURCE,
   VAR_FIELDS,
@@ -33,11 +32,12 @@ import {
   VAR_PATTERNS,
   explorationDS,
 } from 'services/variables';
-import { ServiceSelectionComponent } from './ServiceSelectionComponent';
 
 import pluginJson from '../../plugin.json';
-import { StartingPointSelectedEvent } from 'Components/Forms/SelectFieldButton';
-import { Pattern } from './ServiceSceneTabs/Patterns/Pattern';
+
+import { Pattern } from './Pattern';
+import { ServiceScene } from '../ServiceScene/ServiceScene';
+import { ServiceSelectionComponent, StartingPointSelectedEvent } from '../ServiceSelectionScene/ServiceSelectionScene';
 
 type LogExplorationMode = 'start' | 'logs';
