+++ src/Components/ServiceScene/ShareExplorationButton.tsx
@@ -4,9 +4,9 @@ import { useLocation } from 'react-use';
 import { ToolbarButton } from '@grafana/ui';
 
 import { config } from '@grafana/runtime';
-import { IndexScene } from 'Components/MainScenes/IndexScene';
 import { getUrlForExploration } from 'services/scenes';
 import { copyText } from 'services/text';
+import { IndexScene } from 'Components/IndexScene/IndexScene';
 
 interface ShareExplorationButtonState {
   exploration: IndexScene;
